+++ src/memory/mbc/mbc1.hpp
@@ -58,12 +58,12 @@ class MBC1 : public MemoryBankController
     /**
      * The size of a ROM bank in bytes
      */
-    static const int ROM_BANK_SIZE_IN_BYTES = 16_KiB;
+    static const size_t ROM_BANK_SIZE_IN_BYTES = 16_KiB;
 
     /**
      * The size of a RAM bank in bytes
      */
-    static const int RAM_BANK_SIZE_IN_BYTES = 8_KiB;
+    static const size_t RAM_BANK_SIZE_IN_BYTES = 8_KiB;
 
     /**
      * The id of the ROM bank currently selected.
